@@ -5,42 +5,37 @@ version: 0.1.1+1001
 
 environment:
   sdk: ">=2.1.0 <3.0.0"
-  flutter: "1.13.6"
+  flutter: "1.26.0-17.3.pre"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-
-  intl: ^0.16.0
-  flutter_sheet_localization: ^1.0.0
-  font_awesome_flutter: ^8.7.0
-  provider: ^4.0.4
-  shared_preferences: ^0.5.6+2
-  sembast: ^2.3.0
-  sembast_web: ^1.0.0
-
-  flutter_speed_dial: ^1.2.5
-  smooth_page_indicator: ^0.1.4
-
-  path_provider: ^1.6.5
+  flutter_speed_dial: ^2.1.1
+  font_awesome_flutter: ^9.0.0-nullsafety
+  intl: ^0.17.0-nullsafety.2
   package_info_wrapper:
     git: https://gitlab.com/kitProject/package_info_wrapper.git
+  path_provider: ^2.0.0-nullsafety
+  provider: ^5.0.0-nullsafety.3
+  sembast: ^3.0.0-nullsafety.4
+  sembast_web: ^2.0.0-nullsafety.1
+  shared_preferences: ^2.0.0-nullsafety
+  smooth_page_indicator: ^0.2.0
 
 dev_dependencies:
+  build_runner: ^1.11.1
+  flutter_launcher_icons: ^0.8.1
   flutter_test:
     sdk: flutter
-
-  pedantic: ^1.9.0
-  flutter_sheet_localization_generator: ^1.2.0
-  build_runner: ^1.8.0
-  flutter_launcher_icons: ^0.7.4
+  pedantic: ^1.10.0
 
 dependency_overrides:
-  pedantic: ^1.9.0
+  pedantic: ^1.10.0
 
 flutter:
+  generate: true
   uses-material-design: true
   assets:
     - assets/images/
