@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:idea_note/config/assets.dart';
 import 'package:idea_note/config/route.dart';
-import 'package:idea_note/localization.dart';
 import 'package:idea_note/model/settings_model.dart';
 import 'package:provider/provider.dart';
 
@@ -29,7 +29,7 @@ class Tutorial1Page extends StatelessWidget {
                 padding: const EdgeInsets.only(right: 5, left: 5),
                 child: Text(
                   AppLocalizations.of(context).tutorialText1,
-                  style: themeData.primaryTextTheme.body1
+                  style: themeData.primaryTextTheme.bodyText2
                       .copyWith(color: Colors.white),
                 ),
               ),
@@ -39,7 +39,7 @@ class Tutorial1Page extends StatelessWidget {
             padding: const EdgeInsets.only(bottom: 20, right: 20, left: 20),
             child: Row(
               children: <Widget>[
-                FlatButton(
+                TextButton(
                   child: Text(
                     AppLocalizations.of(context).skip,
                     style: themeData.textTheme.button
@@ -52,7 +52,7 @@ class Tutorial1Page extends StatelessWidget {
                   },
                 ),
                 const Spacer(),
-                FlatButton(
+                TextButton(
                   child: Text(
                     AppLocalizations.of(context).next,
                     style: themeData.textTheme.button
