@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:idea_note/config/assets.dart';
-import 'package:idea_note/localization.dart';
 
 class Tutorial2Page extends StatelessWidget {
   const Tutorial2Page(this._pageController);
@@ -25,7 +25,7 @@ class Tutorial2Page extends StatelessWidget {
                 padding: const EdgeInsets.only(right: 5, left: 5),
                 child: Text(
                   AppLocalizations.of(context).tutorialText2,
-                  style: themeData.primaryTextTheme.body1
+                  style: themeData.primaryTextTheme.bodyText2
                       .copyWith(color: Colors.white),
                 ),
               ),
@@ -36,7 +36,7 @@ class Tutorial2Page extends StatelessWidget {
             child: Row(
               children: <Widget>[
                 const Spacer(),
-                FlatButton(
+                TextButton(
                   child: Text(
                     AppLocalizations.of(context).next,
                     style: themeData.textTheme.button
