@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:idea_note/config/assets.dart';
 import 'package:idea_note/config/route.dart';
-import 'package:idea_note/localization.dart';
 
 class Tutorial3Page extends StatelessWidget {
   const Tutorial3Page(this._pageController);
@@ -26,16 +26,18 @@ class Tutorial3Page extends StatelessWidget {
                 padding: const EdgeInsets.only(right: 5, left: 5),
                 child: Text(
                   AppLocalizations.of(context).tutorialText3,
-                  style: themeData.primaryTextTheme.body1
+                  style: themeData.primaryTextTheme.bodyText2
                       .copyWith(color: Colors.white),
                 ),
               ),
             ),
           ),
-          RaisedButton(
+          ElevatedButton(
             child: Text(AppLocalizations.of(context).doSetting),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(3),
+            style: ElevatedButton.styleFrom(
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(3),
+              ),
             ),
             onPressed: () async {
               await Navigator.of(context).pushNamed(RoutePath.settingTitle);
@@ -46,7 +48,7 @@ class Tutorial3Page extends StatelessWidget {
             child: Row(
               children: <Widget>[
                 const Spacer(),
-                FlatButton(
+                TextButton(
                   child: Text(
                     AppLocalizations.of(context).next,
                     style: themeData.textTheme.button
