@@ -11,7 +11,7 @@ class SettingTitleSingleButton extends StatelessWidget {
     final themeData = Theme.of(context);
     return Padding(
       padding: const EdgeInsets.only(top: 20, bottom: 20),
-      child: OutlineButton(
+      child: OutlinedButton(
         child: Text(_text, style: themeData.textTheme.button),
         onPressed: _onPressed,
       ),
