@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:idea_note/config/route.dart';
 import 'package:idea_note/entity/note.dart';
-import 'package:idea_note/localization.dart';
 import 'package:idea_note/model/home_model.dart';
 import 'package:provider/provider.dart';
 
