@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class MyAppBar extends StatelessWidget implements PreferredSizeWidget {
-  MyAppBar(this._title, {this.automaticallyImplyLeading = true})
-      : preferredSize = Size.fromHeight(kToolbarHeight);
+  const MyAppBar(this._title, {this.automaticallyImplyLeading = true})
+      : preferredSize = const Size.fromHeight(kToolbarHeight);
 
   final String _title;
   final bool automaticallyImplyLeading;
